Make toolbarHeight return a CGFloat.

+++ Custom Windows/DTWindowCategory.m
@@ -10,7 +10,7 @@
 
 @implementation NSWindow (DTWindowExtension)
 
-- (float)toolbarHeight
+- (CGFloat)toolbarHeight
 {
     return NSHeight([self frame]) - NSHeight([[self contentView] frame]);
 }
